+++ buildSrc/src/main/kotlin/versions.kt
@@ -1,13 +1,14 @@
 object vers {
-    val kotlin = "1.4.20"
+    val kotlin = "1.4.21"
     val nexus_staging = "0.22.0"
 
     object kotlinx {
         val datetime = "0.1.1"
     }
+
     object asoft {
-        val logging = "0.0.1"
-        val builders = "1.1.1"
-        val test = "1.0.1"
+        val logging = "0.0.20"
+        val builders = "1.3.0"
+        val expect = "0.0.20"
     }
 }