+++ logging-console/src/commonMain/kotlin/logging/ConsoleAppender.kt
@@ -1,3 +1,3 @@
-package tz.co.asoft
+package logging
 
 expect class ConsoleAppender(options: ConsoleAppenderOptions) : Appender