+++ logging-core/src/commonMain/kotlin/logging/Logging.kt
@@ -1,4 +1,4 @@
-package tz.co.asoft
+package logging
 
 object Logging {
     internal val appenders = mutableListOf<Appender>()