+++ CHANGELOG.md
@@ -1,15 +1,27 @@
-# 0.0.1
+# 0.0.20
+
+- Updated gradle from `6.7.1` to `6.8.1`
+- Changed package name from `tz.co.asoft` to `logging`
+- Replaced dependency of test with expect
+
+# 0.0.10
+
 ## Availability
+
 - Published to maven central
 
 ## Build Src
+
 - Updated to gradle version 6.7.1
 
 ## New Features
+
 - Initial release
 
 ## Documentation
+
 - Update readme
 
 ## Samples
+
 - Added samples